Fixed: Decimal Values in Dialog Boxes

+++ JaVisteAMinhaMedia/app/src/main/java/pt/isec/gps1819g11/javisteaminhamedia/DialogInputs/AverageGradeDialog.java
@@ -68,7 +68,7 @@ public class AverageGradeDialog extends DialogFragment implements View.OnClickLi
 
                 String value = String.valueOf(inputGrade.getText());
                 if(!value.isEmpty()){
-                    float intendedAverage = Integer.parseInt(value);
+                    float intendedAverage = Float.parseFloat(value);
                     if(intendedAverage <= 20 && intendedAverage > 9.5 ){
                         try{
                         student.setIntendedAverage(intendedAverage);
